Added more info about other asserted macros
Added information about:

* unimplemented
* checkCode
* checkNoEntry
* checkNoReentry
* checkNoRecursion

Personal notes:

Hey, sorry for the slow progress with updates. I have the lot of things on my mind. I currently study for a university and has a lot of assignments, as well as working on my own side projects.

So, keep that in mind, that my personal projects will take time to update and fix. Otherwise, I hope you have a wonderful day and enjoy coding!